pccommon/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --extra=server --output-file=pccommon/requirements.txt ./pccommon/setup.py
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via starlette
async-timeout==4.0.3
    # via redis
azure-core==1.30.2
    # via
    #   azure-data-tables
    #   azure-identity
    #   azure-storage-blob
    #   opencensus-ext-azure
azure-data-tables==12.5.0
    # via pccommon (pccommon/setup.py)
azure-identity==1.16.1
    # via
    #   opencensus-ext-azure
    #   pccommon (pccommon/setup.py)
azure-storage-blob==12.20.0
    # via pccommon (pccommon/setup.py)
beautifulsoup4==4.12.3
    # via html-sanitizer
cachetools==5.3.3
    # via
    #   google-auth
    #   pccommon (pccommon/setup.py)
certifi==2024.6.2
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==42.0.8
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
exceptiongroup==1.2.1
    # via anyio
fastapi-slim==0.111.0
    # via pccommon (pccommon/setup.py)
google-api-core==2.19.0
    # via opencensus
google-auth==2.30.0
    # via google-api-core
googleapis-common-protos==1.63.1
    # via google-api-core
html-sanitizer==2.4.4
    # via pccommon (pccommon/setup.py)
idna==3.7
    # via
    #   anyio
    #   pccommon (pccommon/setup.py)
    #   requests
    #   yarl
isodate==0.6.1
    # via
    #   azure-data-tables
    #   azure-storage-blob
lxml==5.2.2
    # via
    #   html-sanitizer
    #   lxml-html-clean
lxml-html-clean==0.1.0
    # via
    #   html-sanitizer
    #   pccommon (pccommon/setup.py)
msal==1.28.1
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.1.0
    # via azure-identity
multidict==6.0.5
    # via yarl
opencensus==0.11.4
    # via
    #   opencensus-ext-azure
    #   opencensus-ext-logging
opencensus-context==0.1.3
    # via opencensus
opencensus-ext-azure==1.1.13
    # via pccommon (pccommon/setup.py)
opencensus-ext-logging==0.1.1
    # via pccommon (pccommon/setup.py)
orjson==3.10.5
    # via pccommon (pccommon/setup.py)
packaging==24.1
    # via msal-extensions
portalocker==2.8.2
    # via msal-extensions
proto-plus==1.23.0
    # via google-api-core
protobuf==4.25.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psutil==5.9.8
    # via opencensus-ext-azure
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.7.4
    # via
    #   fastapi-slim
    #   pccommon (pccommon/setup.py)
    #   pydantic-settings
pydantic-core==2.18.4
    # via pydantic
pydantic-settings==2.3.3
    # via pccommon (pccommon/setup.py)
pyhumps==3.5.3
    # via pccommon (pccommon/setup.py)
pyjwt[crypto]==2.8.0
    # via msal
python-dotenv==1.0.1
    # via pydantic-settings
redis==4.6.0
    # via pccommon (pccommon/setup.py)
requests==2.32.3
    # via
    #   azure-core
    #   google-api-core
    #   msal
    #   opencensus-ext-azure
    #   pccommon (pccommon/setup.py)
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   azure-core
    #   isodate
    #   opencensus
sniffio==1.3.1
    # via anyio
soupsieve==2.5
    # via beautifulsoup4
starlette==0.37.2
    # via
    #   fastapi-slim
    #   pccommon (pccommon/setup.py)
types-cachetools==4.2.9
    # via pccommon (pccommon/setup.py)
typing-extensions==4.12.2
    # via
    #   anyio
    #   azure-core
    #   azure-data-tables
    #   azure-storage-blob
    #   fastapi-slim
    #   pydantic
    #   pydantic-core
    #   starlette
urllib3==2.2.2
    # via
    #   pccommon (pccommon/setup.py)
    #   requests
yarl==1.9.4
    # via azure-data-tables
